+++ packages/alpha/src/api/alerts/subscriptionsApi.ts
@@ -1,8 +1,11 @@
+// Copyright 2022 Cognite AS
+
 import { BaseResourceAPI, CDFHttpClient, MetadataMap } from '@cognite/sdk-core';
 import {
   Subscription,
   SubscriptionCreate,
   SubscriptionDelete,
+  SubscriptionFilterQuery,
 } from '../../types';
 
 export class SubscriptionsAPI extends BaseResourceAPI<Subscription> {
@@ -16,5 +19,12 @@ export class SubscriptionsAPI extends BaseResourceAPI<Subscription> {
 
   public delete = async (items: SubscriptionDelete[]) => {
     return this.deleteEndpoint(items);
+  };
+
+  public list = async (filter?: SubscriptionFilterQuery) => {
+    return this.listEndpoint<SubscriptionFilterQuery>(
+      this.callListEndpointWithPost,
+      filter
+    );
   };
 }