+++ packages/alpha/src/types.ts
@@ -17,9 +17,84 @@ import {
 
 export * from '@cognite/sdk';
 
-export interface AlertFilter extends FilterQuery {
-  channelId?: number;
-  channelExternalId?: string;
+export type MonitoringTaskModelExternalId = 'threshold';
+
+export const MonitoringTaskModelExternalId = {
+  THRESHOLD: 'threshold' as MonitoringTaskModelExternalId,
+};
+
+export interface MonitoringTaskThresholdModelCreate {
+  externalId: MonitoringTaskModelExternalId;
+  timeseriesExternalId: CogniteExternalId;
+  threshold: number;
+  granularity?: string;
+}
+
+export interface MonitoringTaskThresholdModel {
+  externalId: MonitoringTaskModelExternalId;
+  timeseriesId: CogniteInternalId;
+  granularity?: string;
+  threshold: number;
+}
+
+export interface MonitoringTaskCreate {
+  externalId: CogniteExternalId;
+  channelId: CogniteInternalId;
+  interval: number;
+  overlap: number;
+  model: MonitoringTaskThresholdModelCreate;
+  nonce: string;
+}
+
+export interface MonitoringTask {
+  id: CogniteInternalId;
+  externalId?: CogniteExternalId;
+  channelId: CogniteInternalId;
+  interval: number;
+  overlap: number;
+  model: MonitoringTaskThresholdModel;
+}
+
+export interface MonitoringTaskModelFilter {
+  timeseriesId: CogniteInternalId;
+}
+
+export interface MonitoringTaskFilter {
+  externalIds?: CogniteExternalId[];
+  ids?: CogniteInternalId[];
+  channelIds?: CogniteInternalId[];
+  model?: MonitoringTaskModelFilter;
+}
+
+export interface MonitoringTaskFilterQuery extends FilterQuery {
+  filter?: MonitoringTaskFilter;
+}
+
+export interface AlertFilter {
+  channelIds?: CogniteInternalId[];
+  channelExternalIds?: CogniteExternalId[];
+}
+
+export interface AlertFilterQuery extends FilterQuery {
+  filter?: AlertFilter;
+}
+
+export interface AlertDeduplicationRuleCreate {
+  mergeInterval?: string;
+  activationInterval?: string;
+}
+
+export interface AlertRulesCreate {
+  deduplication: AlertDeduplicationRuleCreate;
+}
+
+export interface AlertDeduplicationRule {
+  mergeInterval?: string;
+  activationInterval?: string;
+}
+
+export interface AlertRules {
+  deduplication: AlertDeduplicationRule;
 }
 
 export interface AlertCreate {
@@ -54,6 +129,7 @@ export interface ChannelCreate {
   name: string;
   description?: string;
   metadata?: Metadata;
+  alertRules?: AlertRulesCreate;
 }
 
 export interface Channel {
@@ -64,13 +140,18 @@ export interface Channel {
   name: string;
   description?: string;
   metadata?: Metadata;
+  alertRules?: AlertRules;
 }
 
-export interface ChannelFilter extends FilterQuery {
+export interface ChannelFilter {
   ids?: CogniteInternalId[];
   externalIds?: CogniteExternalId[];
   parentIds?: CogniteInternalId[];
   metadata?: Metadata;
+}
+
+export interface ChannelFilterQuery extends FilterQuery {
+  filter?: ChannelFilter;
 }
 
 export interface ChannelPatch {
@@ -90,6 +171,17 @@ export interface SubscriberCreate {
   externalId?: CogniteExternalId;
   metadata?: Metadata;
   email: string;
+}
+
+export interface SubscriberFilter {
+  externalIds?: CogniteExternalId[];
+  ids?: CogniteInternalId[];
+  metadata?: Metadata;
+  email?: string;
+}
+
+export interface SubscriberFilterQuery extends FilterQuery {
+  filter?: SubscriberFilter;
 }
 
 export interface Subscriber {
@@ -112,6 +204,19 @@ export interface SubscriptionCreate {
   subscriberId?: CogniteInternalId;
   subscriberExternalId?: CogniteExternalId;
   metadata?: Metadata;
+}
+
+export interface SubscriptionFilter {
+  externalIds?: CogniteExternalId[];
+  channelIds?: CogniteInternalId[];
+  channelExternalIds?: CogniteExternalId[];
+  subscriberIds?: CogniteInternalId[];
+  subscriberExternalIds?: CogniteExternalId[];
+  metadata?: Metadata;
+}
+
+export interface SubscriptionFilterQuery extends FilterQuery {
+  filter?: SubscriptionFilter;
 }
 
 export interface SubscriptionDelete {