+++ packages/codegen/src/__tests__/codegen.uint.spec.ts
@@ -2,6 +2,8 @@ import {
   CodeGen,
   passThroughFilter,
   createServiceNameFilter,
+  createPathFilter,
+  StringFilter,
 } from '../codegen';
 import { OpenApiDocument } from '../openapi';
 import { OpenApiSnapshotManager } from '../snapshot';
@@ -80,6 +82,57 @@ describe('code generation', () => {
       const paths = gen['filterPaths'](basicSnapshot.paths);
       expect(Object.keys(paths)).toHaveLength(4);
     });
+
+    test('ignore filter', async () => {
+      const filter = (
+        data: Record<string, object>,
+        predicate: StringFilter
+      ): object => {
+        return Object.keys(data)
+          .filter(predicate)
+          .reduce((acc, path) => {
+            return Object.assign(acc, { [path]: data[path] });
+          }, {});
+      };
+
+      const base = '/api/v1/projects/{project}';
+      const data = {
+        [`${base}/serviceA`]: {},
+        [`${base}/serviceB`]: {},
+        [`${base}/serviceA/serviceC`]: {},
+        [`${base}/serviceA/serviceD`]: {},
+      };
+
+      const predicateServiceA = createPathFilter(
+        [createServiceNameFilter('serviceA')],
+        []
+      );
+      expect(Object.keys(filter(data, predicateServiceA))).toEqual([
+        `${base}/serviceA`,
+        `${base}/serviceA/serviceC`,
+        `${base}/serviceA/serviceD`,
+      ]);
+
+      const predicateServiceAAndC = createPathFilter(
+        [createServiceNameFilter('serviceA')],
+        [createServiceNameFilter('serviceA/serviceD')]
+      );
+      expect(Object.keys(filter(data, predicateServiceAAndC))).toEqual([
+        `${base}/serviceA`,
+        `${base}/serviceA/serviceC`,
+      ]);
+
+      const predicateServiceAOnly = createPathFilter(
+        [createServiceNameFilter('serviceA')],
+        [
+          createServiceNameFilter('serviceA/serviceC'),
+          createServiceNameFilter('serviceA/serviceD'),
+        ]
+      );
+      expect(Object.keys(filter(data, predicateServiceAOnly))).toEqual([
+        `${base}/serviceA`,
+      ]);
+    });
   });
 
   describe('generate types', () => {
@@ -156,6 +209,39 @@ describe('code generation', () => {
         await fs.readFile(CodeGen.outputFileName)
       ).toString();
       expect(generatedFile).toEqual(cyclicReferencesGenFile);
+    });
+
+    test('restrict to relevant definitions', async () => {
+      const snapshotMngr = new OpenApiSnapshotManager({ directory: '.' });
+      const snapshot = await snapshotMngr.downloadFromPath({
+        path: testFolder + '/testdata',
+        filename: '8-paths.json',
+      });
+
+      const gen = new CodeGen(new AcacodeOpenApiGenerator(), {
+        autoNameInlinedRequest: false,
+        outputDir: process.cwd(),
+        filter: {
+          path: passThroughFilter,
+        },
+      });
+
+      const typeNames = await gen.generateTypes(snapshot, [
+        'FunctionListScope',
+      ]);
+
+      expect(typeNames).toEqual([
+        'CogniteExternalId',
+        'CogniteInternalId',
+        'EpochTimestamp',
+        'FunctionFileId',
+        'FunctionFilter',
+        'FunctionListScope',
+        'FunctionName',
+        'FunctionOwner',
+        'FunctionStatus',
+        'LimitList',
+      ]);
     });
   });
 });