+++ packages/playground/src/__tests__/testFunction/handler.py
@@ -1,3 +1,0 @@
-def handle(data):
-    print('Hello World!')
-    print(f'Received data: {data}')