+++ packages/stable/src/types.ts
@@ -10,6 +10,7 @@ import {
   InternalId,
   Limit,
 } from '@cognite/sdk-core';
+import { AnnotationData } from './api/annotations/types.gen';
 
 export {
   ListResponse,
@@ -1462,16 +1463,25 @@ export interface List3DNodesQuery extends FilterQuery {
   sortByNodeId?: boolean;
 }
 
+export interface Filter3DNodesByNames {
+  /**
+   * Name filter. Nodes satisfy the filter if, the name of the node exists in the provided array.
+   */
+  names: string[];
+}
+
+export interface Filter3DNodesByProperty {
+  /**
+   * Property filters. Nodes satisfy the filter if, for each property in the nested map(s), they have a value corresponding to that property that is contained within the list associated with that property in the map.
+   */
+  properties: { [key: string]: { [key: string]: string[] } };
+}
+
 export interface Filter3DNodesQuery extends FilterQuery {
   /**
    * List filter
    */
-  filter: {
-    /**
-     * Property filters. Nodes satisfy the filter if, for each property in the nested map(s), they have a value corresponding to that property that is contained within the list associated with that property in the map.
-     */
-    properties: { [key: string]: { [key: string]: string[] } };
-  };
+  filter: Filter3DNodesByProperty | Filter3DNodesByNames;
   /**
    * Partition specifier of the form "n/m". It will return the n'th (1-indexed) part of the result divided into m parts.
    */
@@ -1989,6 +1999,16 @@ export interface Revision3D {
    * The rotation is expressed by Euler angles in radians and in XYZ order.
    */
   rotation?: Tuple3<number>;
+  /**
+   * Global translation to be applied to the entire model.
+   * The translation is expressed in meters.
+   */
+  translation?: Tuple3<number>;
+  /**
+   * Global scale to be applied to the entire model.
+   * The scale is expressed as scale along the X, Y and Z axes.
+   */
+  scale?: Tuple3<number>;
   camera?: RevisionCameraProperties;
   /**
    * The status of the revision.
@@ -2524,6 +2544,16 @@ export interface UpdateRevision3D {
      * The rotation is expressed by Euler angles in radians and in XYZ order.
      */
     rotation?: SetField<Tuple3<number>>;
+    /**
+     * Global translation to be applied to the entire model.
+     * The translation is expressed in meters.
+     */
+    translation?: SetField<Tuple3<number>>;
+    /**
+     * Global scale to be applied to the entire model.
+     * The scale is expressed as scale along the X, Y and Z axes.
+     */
+    scale?: SetField<Tuple3<number>>;
     /**
      * Initial camera target.
      */
@@ -3208,12 +3238,11 @@ export interface RelationshipsRetrieveParams extends IgnoreUnknownIds {
   fetchResources?: boolean;
 }
 
-export type AnnotatedResourceType = 'file';
+export type AnnotatedResourceType = 'file' | 'threedmodel';
 export type AnnotationStatus = 'suggested' | 'approved' | 'rejected';
 
 // TODO [CXT-463] Use annotation-types package definitions
 export type AnnotationType = string;
-export type AnnotationPayload = object;
 
 export interface AnnotationModel extends AnnotationCreate {
   id: CogniteInternalId;
@@ -3232,7 +3261,7 @@ export interface AnnotationSuggest {
   creatingApp: string;
   creatingAppVersion: string;
   creatingUser: string | null;
-  data: AnnotationPayload;
+  data: AnnotationData;
 }
 
 export interface AnnotationChangeById extends InternalId, AnnotationUpdate {}
@@ -3240,7 +3269,7 @@ export interface AnnotationChangeById extends InternalId, AnnotationUpdate {}
 export interface AnnotationUpdate {
   update: {
     annotationType?: SetField<AnnotationType>;
-    data?: SetField<AnnotationPayload>;
+    data?: SetField<AnnotationData>;
     status?: SetField<AnnotationStatus>;
   };
 }
@@ -3260,5 +3289,5 @@ export interface AnnotationFilterProps {
   creatingAppVersion?: string;
   creatingUser?: string | null;
   status?: AnnotationStatus;
-  data?: AnnotationPayload;
+  data?: Partial<AnnotationData>;
 }