+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.25](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.24...@cognite/sdk-beta@5.2.25) (2023-01-24)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.24](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.23...@cognite/sdk-beta@5.2.24) (2023-01-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.23](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.22...@cognite/sdk-beta@5.2.23) (2023-01-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.22](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.21...@cognite/sdk-beta@5.2.22) (2023-01-09)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.21](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.20...@cognite/sdk-beta@5.2.21) (2023-01-06)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.20](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.19...@cognite/sdk-beta@5.2.20) (2022-12-09)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.19](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.18...@cognite/sdk-beta@5.2.19) (2022-11-21)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.18](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.17...@cognite/sdk-beta@5.2.18) (2022-11-18)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.17](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.16...@cognite/sdk-beta@5.2.17) (2022-11-17)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.16](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.15...@cognite/sdk-beta@5.2.16) (2022-11-02)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.2.15](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.14...@cognite/sdk-beta@5.2.15) (2022-10-28)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
 ## [5.2.14](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.13...@cognite/sdk-beta@5.2.14) (2022-10-04)
 
 **Note:** Version bump only for package @cognite/sdk-beta